Updating the "main" to point to the built JS
This makes bundling via Browserify a TON easier than the coffee path that main.js takes.  Plus, if you are delivering the JS, might as well have consumers use it directly.

While I was in there, I added the "repository" field so that npmjs.org will link to the source properly when the module gets published.